Fix errors in D3 file, and change EventTarget to Element
== EventTarget -> Element ==
Instead of an EventTarget , node() should return Element,
and the various select() variants should also take in  Element.

Commit d9cb8491 Changed the type to EventTarget, for no listed reason.
Inspecting the code, d3.js uses querySelector() as part of the select()
call, which returns Element, so that should be the type.

== Interval.offset() ==
Should take a number not a Date as the second parameter
(https://github.com/mbostock/d3/wiki/Time-Intervals#interval_offset)

== svg.line.defined() and svg.area.defined() ==
In two places setter signature for defined() wasn't listed.

== svg.symbol()'s index parameter is optional ==
https://github.com/mbostock/d3/wiki/SVG-Shapes#_symbol